requirements.txt
google-play-scraper
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
emoji
konlpy
JPype1
transformers[torch]
torch
numpy